+++ web/src/main.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import Router from "./components/router";
-import UserAuthProvider from "./context/auth/provider";
+import UserAuthProvider from "./store/auth/provider";
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
